+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 ---
 ### 28 MAR 2025
+- Updated all `onclick="closeDialog()"` to `onclick="closeDialog(event)"` to accommodate `closeDialog` function scope changes.
+- The `closeDialog` function `dialogController.js` module now checks the `tagName` property for clicked element.
+- Removed `stopPropagation` function in `dialogController.js` module.
 - Created `dialog-controller` branch from `styles` branch.
 - Added `onclick` attribute to `fileDetails.ejs` partial to stop event propagation.
 - Inputs only have a `bottom-border` property.